The difference between two pointers can be long. There is no need to increase the version number here because the IO implementation is oblivious to the length of the integer types.

+++ contrib/mul/vil3d/io/vil3d_io_image_view.h
@@ -27,7 +27,7 @@ inline void vsl_b_write(vsl_b_ostream &os, const vil3d_image_view<T>& image)
   {
     vsl_b_write(os, image.memory_chunk());
 
-    int offset = (image.origin_ptr()-(const T*)image.memory_chunk()->data());
+    long offset = (image.origin_ptr()-(const T*)image.memory_chunk()->data());
     vsl_b_write(os, offset);
   }
 }
@@ -42,7 +42,7 @@ inline void vsl_b_read(vsl_b_istream &is, vil3d_image_view<T>& image)
   unsigned ni,nj,nk,np;
   int istep,jstep,kstep,pstep;
   vil2_memory_chunk_sptr chunk;
-  int offset;
+  long offset;
 
   short w;
   vsl_b_read(is, w);
